+++ src/test/java/org/fugerit/java/junit5/tag/check/PrepareTest.java
@@ -1,0 +1,12 @@
+package org.fugerit.java.junit5.tag.check;
+
+import org.junit.jupiter.api.Order;
+
+@Order(1)
+class PrepareTest {
+
+    void test1() {
+
+    }
+
+}