+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- documented output in json, xml and text
+
 ## [1.0.0] - 2025-11-26
 
 ### Added